+++ app/src/main/java/com/meli/android/carddrawer/app/MainActivity.java
@@ -13,6 +13,7 @@ import android.widget.Switch;
 import android.widget.TextView;
 import com.meli.android.carddrawer.app.model.CardComposite;
 import com.meli.android.carddrawer.app.model.MasterCardConfiguration;
+import com.meli.android.carddrawer.app.model.UrlTestConfiguration;
 import com.meli.android.carddrawer.app.model.VisaCardBlueConfiguration;
 import com.meli.android.carddrawer.app.model.VisaCardGrayConfiguration;
 import com.meli.android.carddrawer.app.model.VisaCardGreenConfiguration;
@@ -28,6 +29,7 @@ public class MainActivity extends AppCompatActivity {
     private CardComposite card;
     private CardDrawerView cardDrawerView;
     private CardDrawerView cardDrawerViewLowRes;
+    private CardDrawerView cardDrawerViewMedium;
 
     @Override
     protected void onCreate(final Bundle savedInstanceState) {
@@ -36,9 +38,11 @@ public class MainActivity extends AppCompatActivity {
 
         cardDrawerView = findViewById(R.id.card_header_container);
         cardDrawerViewLowRes = findViewById(R.id.card_header_lowres_container);
+        cardDrawerViewMedium = findViewById(R.id.card_header_medium_container);
         card = new CardComposite();
         card.addCard(cardDrawerView.getCard());
         card.addCard(cardDrawerViewLowRes.getCard());
+        card.addCard(cardDrawerViewMedium.getCard());
         ((Switch) findViewById(R.id.card_header_switch_responsive)).setOnCheckedChangeListener(
             new CompoundButton.OnCheckedChangeListener() {
                 @Override
@@ -47,6 +51,7 @@ public class MainActivity extends AppCompatActivity {
                         CardDrawerView.Behaviour.RESPONSIVE : CardDrawerView.Behaviour.REGULAR;
                     cardDrawerView.setBehaviour(behaviour);
                     cardDrawerViewLowRes.setBehaviour(behaviour);
+                    cardDrawerViewMedium.setBehaviour(behaviour);
                 }
             });
         ((Switch) findViewById(R.id.card_header_lowres_switch)).setOnCheckedChangeListener(
@@ -57,6 +62,15 @@ public class MainActivity extends AppCompatActivity {
                     cardDrawerView.setVisibility(isChecked ? View.GONE : View.VISIBLE);
                 }
             });
+        ((Switch) findViewById(R.id.card_header_medium_switch)).setOnCheckedChangeListener(
+                new CompoundButton.OnCheckedChangeListener() {
+                    @Override
+                    public void onCheckedChanged(final CompoundButton buttonView, final boolean isChecked) {
+                        cardDrawerViewMedium.setVisibility(isChecked ? View.VISIBLE : View.GONE);
+                        cardDrawerView.setVisibility(isChecked ? View.GONE : View.VISIBLE);
+                    }
+                });
+
         initCardConfigurationOptions();
         initCardNumber();
         initCardName();
@@ -72,6 +86,7 @@ public class MainActivity extends AppCompatActivity {
             public void beforeTextChanged(CharSequence charSequence, int i, int i1, int i2) {
                 cardDrawerView.showSecurityCode();
                 cardDrawerViewLowRes.showSecurityCode();
+                cardDrawerViewMedium.showSecurityCode();
             }
 
             @Override
@@ -84,6 +99,7 @@ public class MainActivity extends AppCompatActivity {
                 if (editable.toString().isEmpty()) {
                     cardDrawerView.show();
                     cardDrawerViewLowRes.show();
+                    cardDrawerViewMedium.show();
                 }
             }
         });
@@ -135,6 +151,7 @@ public class MainActivity extends AppCompatActivity {
             public void onClick(View v) {
                 cardDrawerView.show();
                 cardDrawerViewLowRes.show();
+                cardDrawerViewMedium.show();
             }
         });
     }
@@ -164,6 +181,7 @@ public class MainActivity extends AppCompatActivity {
             public void onClick(View v) {
                 cardDrawerView.show();
                 cardDrawerViewLowRes.show();
+                cardDrawerViewMedium.show();
             }
         });
     }
@@ -194,6 +212,9 @@ public class MainActivity extends AppCompatActivity {
         MasterCardConfiguration masterCardConfiguration = new MasterCardConfiguration(this);
         cardOptions.add(new CardConfigurationOption(masterCardConfiguration, "Master"));
 
+        final UrlTestConfiguration urlTestConfiguration = new UrlTestConfiguration(this);
+        cardOptions.add(new CardConfigurationOption(urlTestConfiguration, "Url Test"));
+
         ArrayAdapter<CardConfigurationOption> cardAdapter = new ArrayAdapter(this, android.R.layout.simple_spinner_item, cardOptions);
         cardAdapter.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item);
         cardsSpinner.setAdapter(cardAdapter);
@@ -203,6 +224,7 @@ public class MainActivity extends AppCompatActivity {
                 CardConfigurationOption selection = cardOptions.get(position);
                 cardDrawerView.show(selection.getCardConfiguration());
                 cardDrawerViewLowRes.show(selection.getCardConfiguration());
+                cardDrawerViewMedium.show(selection.getCardConfiguration());
             }
 
             @Override