+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+### Release 2.0.2
+Back merge problem
+
+### Release 2.0.1
+Fixed arrow color for medium card
+
+### Release 2.0.0
+Migration to api 19
+Add logos from url support
+
+### Release 1.2.1
+Fix Height Images Card Medium
+
+### Release 1.2.0
+Add Medium Card View
+Add method to customize overlay
+
+### Release 1.1.0
+Monospace support to api less than 21
+Update lowres
+
 ## [First Release (1.0.0)]
 ### 🚀 First Release 🚀
-Stable 1.0.0 version Android MLCardDrawer.
+Stable 1.0.0 version Android MLCardDrawer